cpf_validator/requirements.txt
asgiref==3.3.4
astroid==2.5.3
colorama==0.4.4
Django==3.2
djangorestframework==3.12.4
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pylint==2.7.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
yapf==0.31.0
